requirements.txt
PyMySQL=0.6.6
PyYAML==3.11
irc==12.4.3
